Code Simplification
Updated/Changed the whole CollectionView datasource code, be easier to
use and accept cells that are to be created in IB and not just via
code, all held together with new protocols and delegates (more may be
added with new functions to the app).

Download Manager has been simplified/updated so users can use
delegation over completion handlers, allows for neater looking code
thats easier to follow, with some new variables inside download
manager, will be updating `init’s`/functions to work with the new
download manager.

Updates to follow will be testing, refactoring and removal of redundant
code for all parts of the app, also changing subclasses around so will
be easier to follow.