Download Manager Update + QRCode fixes
Changed the download manager to conform to a typealias  `Networkable`
which contains all of the original manager protocols, but the new
`updatable` and renamed `FIRDownloadManager`. This reduced the number
of lines needed to enter and allows for an easier understanding of what
the download manager is,  maybe change the name to `NetworkManager`?

Also fixed the QRCode reader/creator so it creates the QRCode with the
base url for the database, this means with a checker the scanner will
only accept QRCodes that have the base url, no issues then with
accidentally scanning random QRCodes. Added the ability for when the
QRCode is scanned the code now updates the database, and re-downloads
the ‘scanned’ node to repopulate the collection view with the new card
(This may change when the ui is implemented as it may be easier to find
a way to place the Post inside the collection view objects without a
full reload)