Delegation Naming Update + HWCollectionView Update fixes  + StickerCollectionViewFlowLayout added+ New Exts
Updated the naming of my AuthManagerDelegate to allow it to feel more
like it belongs in the code.

Also updated the collection view code to conform with the superclass
name changes inside HWCollectionView

Added some helper texts for shadow and UIColor, with the ability to
change the fonts colour depending on the cell background colour, may be
useful later on.

TODO:

Comment more of the code so that if the app does well upon launch and
we required more help they will understand there way around the whole
HWCollectionView datasource and delegation